add RelayObservable
